Add src/helpers/print.py
Function _print() takes a string and displays that string on the terminal.
This function also adds left 'whitespace' padding to the given string if there is a
'pad' property present in the 'kwargs' otherwise it simply displays that string.

Args:
    string: it is the string that needs to be printed on the terminal.
    **kwargs: are the key-value pair that print method takes besides string.

Function printRed() takes a string and displays that string in red color.

Args:
    string: it is the string that needs to be printed on the terminal.
    **kwargs: are the key-value pair that print method takes besides string.

Function printGreen() takes a string and displays that string in red green.

Args:
    string: it is the string that needs to be printed on the terminal.
    **kwargs: are the key-value pair that print method takes besides string.

Function printBlue() takes a string and displays that string in blue color.

Args:
    string: it is the string that needs to be printed on the terminal.
    **kwargs: are the key-value pair that print method takes besides string.

Function printYellow() takes a string and displays that string in yellow color.

Args:
    string: it is the string that needs to be printed on the terminal.
    **kwargs: are the key-value pair that print method takes besides string.