feat (cli): add functions for invitation status
Method calculate_status() sets the value of '_stat' (status) according to the invitation
status which turns out to be a  mark on status 'sent' and a  mark on status 'failed'.

Args:
    - self: class object which has the property 'calculate_status' in it.
    - status: status of the request sent or failed.

return:
    - {String} invitation status.

Method show() shows the invitation status on the terminal window in a specific format
    {Status} Person Name {Normal Style}
    Person Occupation {Dim Style}

    Succes rate: # {Green} Failure rate: # {Red} Elapsed time: # {Blue}

Args:
    - self: class object.
    - obj: from which the name and occupation to be pulled.
    - status: status of the request sent or failed.
    - elapsed_time: how much time did it take to calculate and print the invitation status
        for every person.

Function reset() resets the class static variables to their original values once the
user is done sending the invitations to LinkedIn users.

Destructor resets the class static attributes.