feat (javascript): get javascript from *.js file and run on the dom
Function get_javascript() returns the javascript of a file specified.

This function takes the file path as a parameter and checks if it's a absolute path
or relative to the project's source directory. If the file_path is found to be a
relative path then this function converts that relative file path to absolute.

return:
    - javascript of the file specified.