refactor (dom javascript): separate module for javascript execution
Function get_page_y_offset() returns the window.pageYOffset
of the webpage, we need that so we can keep on scrolling untill
the page offset becomes constant. Declaration of this method is
static because we want to use this function across multiple
classes.

Args:
    self: is not a object here but it is a parameter object
    that has a property 'driver' and we need that

return:
    window.pageYOffset

Function execute_javascript() scrolls the web page to the
very bottom of it using the 'document.scrollingElement.scrollTop'
property.

Args:
    self: it is a parameter object that has a property 'driver'
    in it and we need that to access the webpage.