feat (input): src/helpers/getch()
This functionality of getch() is needed in the input.py file because, taking
input character by character let us examine what command the user is trying
to enter. Once we predicted that we come up with suggestions on the terminal,
basically I want to implement a command completion tool for linkedin-bot.

This implementation of getch() function became more easy with the help of the
user Phylliida who has posted a solution on stackoverflow,
see https://stackoverflow.com/a/31749681/13910122