refactor (input): src/helpers/input.py
separate module for custom _input function.

Function _input() is a dedicated input method for our LinkedIn `cli`
(Command Line Interface), it also handles the Keyboard interrupt error.
We declare this function static because we don't need to give this
function an access to the object it's no use giving this function access
to the object.

return:
   returns the entered value