feat (console): separate module for console messages
we write a different module to enable or disable console messages, if console messages
are disabled then any module file except for main.py file can't print any message on the
screen but if console messages are enabled then any module file can print message on the
screen.

this modification is to make sure that no module file other than main.py file can make
a interface if it is not required.