Separate function `GetElementByXPath()` into a separate utility module
We need this functionality of finding web elements with an explicit wait with
`TimeoutException` handling functionality in two different modules one is
`linkedinconnect` and the other one is `linkedinsearchconnect` so we must make it
re-useable by separating it into a different module.

Also, reduce the default wait time to `10` seconds because `60` is too long.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>