Remove packages that are not in use anymore
Long ago I thought to use pretty complex library to make a pretty robust
automation software but now I've realised that it was a waste of time as I
have to update the elements path inside of my project after each week because
LinkedIn does that.

To make the application simpler remove the implementation of un-neccessary
ideas and install reliable packages to lower the work load.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>